Split tests to integration and unit projects
