+++ src/log4stash.IntegrationTests/Program.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace log4stash.Tests
+{
+    public static class Program
+    {
+        public static void Main()
+        {
+            ElasticsearchAppenderTests.Performance();
+            Console.ReadLine();
+        }
+    }
+}
