+++ src/log4stash.IntegrationTests/app.config
@@ -1,0 +1,18 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+  <runtime>
+    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+      <dependentAssembly>
+        <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-8.5.0.0" newVersion="6.0.4" />
+      </dependentAssembly>
+    </assemblyBinding>
+  </runtime>
+  <system.diagnostics>
+    <trace autoflush="true">
+      <listeners>
+        <add name="consoleListener" type="System.Diagnostics.ConsoleTraceListener" />
+      </listeners>
+    </trace>
+  </system.diagnostics>
+</configuration>
