--- src/log4stash.Tests/Integration/ConnectionTests.cs
@@ -1,50 +1,0 @@
-using System;
-using System.Linq;
-using Nest;
-using NUnit.Framework;
-
-namespace log4stash.Tests.Integration
-{
-    [TestFixture]
-    public class ElasticsearchTests : TestsSetup
-    {
-        [Test]
-        public void Can_insert_record()
-        {
-            var logEvent = new
-            {
-                ClassName = "IntegrationTestClass",
-                Domain = "TestDomain",
-                Exception = "This is a test exception",
-                FileName = "c:\\test\\file.txt",
-                Fix = "none",
-                FullInfo = "A whole bunch of error info dump",
-                Identity = "localhost\\user",
-                Level = "9",
-                LineNumber = "99",
-                TimeStamp = DateTime.Now
-            };
-
-            var results = Client.Index(logEvent, descriptor => descriptor.Type("anonymous"));
-
-            Assert.That(results.Id, Is.Not.Null.Or.Empty);
-        }
-
-        [Test]
-        public void Can_read_inserted_record()
-        {
-            var logEvent = new
-            {
-                ClassName = "IntegrationTestClass",
-                Exception = "ReadingTest"
-            };
-
-            Client.Index(logEvent, descriptor => descriptor.Type("anonymous"));
-            Client.Refresh(Indices.AllIndices);
-
-            var searchResults = Client.Search<dynamic>(s => s.AllTypes().MatchAll());
-            Assert.AreEqual(1, searchResults.Total);
-            Assert.AreEqual("ReadingTest", searchResults.Documents.First().exception.ToString());
-        }
-    }
-}
