--- src/log4stash.Tests/Program.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using log4stash.Tests.Integration;
-
-namespace log4stash.Tests
-{
-    public static class Program
-    {
-        public static void Main()
-        {
-            ElasticsearchAppenderTests.Performance();
-            Console.ReadLine();
-        }
-    }
-}
