--- src/log4stash.Tests/README.txt
@@ -1,5 +1,0 @@
-﻿The log4stash test suite is a set of integration tests, meaning you should
-have an instance of ElasticSearch installed in order to run them.
-
-You can always find the latest version at https://www.elastic.co/downloads/elasticsearch
-
