--- src/log4stash.Tests/packages.config
@@ -1,22 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="4.4.0" targetFramework="net45" />
-  <package id="Elasticsearch.Net" version="5.5.0" targetFramework="net45" />
-  <package id="FluentAssertions" version="5.10.3" targetFramework="net45" />
-  <package id="log4net" version="2.0.8" targetFramework="net45" />
-  <package id="NEST" version="5.5.0" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net45" />
-  <package id="NSubstitute" version="4.2.1" targetFramework="net45" />
-  <package id="NUnit" version="3.8.1" targetFramework="net45" />
-  <package id="NUnit.Console" version="3.7.0" />
-  <package id="NUnit.ConsoleRunner" version="3.7.0" />
-  <package id="NUnit.Extension.NUnitProjectLoader" version="3.5.0" />
-  <package id="NUnit.Extension.NUnitV2Driver" version="3.6.0" />
-  <package id="NUnit.Extension.NUnitV2ResultWriter" version="3.5.0" />
-  <package id="NUnit.Extension.TeamCityEventListener" version="1.0.2" />
-  <package id="NUnit.Extension.VSProjectLoader" version="3.5.0" />
-  <package id="NUnit.Runners" version="3.7.0" />
-  <package id="System.Threading.Tasks.Extensions" version="4.3.0" targetFramework="net45" />
-  <package id="System.ValueTuple" version="4.4.0" targetFramework="net45" />
-  <package id="NUnit3TestAdapter" version="3.16.1" targetFramework="net45" developmentDependency="true" />
-</packages>
