+++ src/log4stash.UnitTests/Program.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace log4stash.UnitTests
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+        }
+    }
+}
