+++ src/log4stash.UnitTests/TemplateInfoTests.cs
@@ -4,7 +4,7 @@ using log4stash.FileAccess;
 using NSubstitute;
 using NUnit.Framework;
 
-namespace log4stash.Tests.Unit
+namespace log4stash.UnitTests
 {
     class TemplateInfoTests
     {
