+++ src/log4stash.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30723.0
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29911.84
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "log4stash", "log4stash\log4stash.csproj", "{22964CB0-8A39-41EA-807D-B5216D27DF70}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "log4stash.Tests", "log4stash.Tests\log4stash.Tests.csproj", "{2F7D7AD4-2DA0-430A-9E90-9D0E398D590C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "log4stash.IntegrationTests", "log4stash.IntegrationTests\log4stash.IntegrationTests.csproj", "{2F7D7AD4-2DA0-430A-9E90-9D0E398D590C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{462F257B-5EDE-4409-9982-AFA7B17BDFF8}"
 	ProjectSection(SolutionItems) = preProject
@@ -28,6 +28,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".ci", ".ci", "{DB38FCDE-D40
 		.ci\RunTests.sh = .ci\RunTests.sh
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "log4stash.UnitTests", "log4stash.UnitTests\log4stash.UnitTests.csproj", "{14A81F04-BB13-490C-9CC0-B9EA8FCD8850}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -46,10 +48,10 @@ Global
 		{055FD962-C4DC-4E3C-BE56-E1E0EAA4ACCA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{055FD962-C4DC-4E3C-BE56-E1E0EAA4ACCA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{055FD962-C4DC-4E3C-BE56-E1E0EAA4ACCA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F9DD323A-1EB7-404A-81FE-EE4FA711A483}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F9DD323A-1EB7-404A-81FE-EE4FA711A483}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F9DD323A-1EB7-404A-81FE-EE4FA711A483}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F9DD323A-1EB7-404A-81FE-EE4FA711A483}.Release|Any CPU.Build.0 = Release|Any CPU
+		{14A81F04-BB13-490C-9CC0-B9EA8FCD8850}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{14A81F04-BB13-490C-9CC0-B9EA8FCD8850}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14A81F04-BB13-490C-9CC0-B9EA8FCD8850}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{14A81F04-BB13-490C-9CC0-B9EA8FCD8850}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
